WIP: Add more python scripts and rename stuff
