+++ .goreleaser.yml
@@ -3,7 +3,7 @@ archives:
       name_template: "{{ .ProjectName }}_{{ .Version }}_{{ .Os }}_{{ .Arch }}"
 before:
     hooks:
-        - sdk_version=$(go list -m github.com/hashicorp/terraform-plugin-sdk/v2 | sed -n -e 's/^.* //p')
+        - sdk_version=$(go list -m -f '{{ .Version }}' github.com/hashicorp/terraform-plugin-sdk/v2)
         - go mod edit --replace github.com/hashicorp/terraform-plugin-sdk/v2=github.com/cloutierMat/terraform-plugin-sdk/v2@$sdk_version
         - go mod tidy
 builds:
